+++ packages/frontend/src/components/message/MessageList.tsx
@@ -985,6 +985,8 @@ function JumpDownButton({
   >['store']['effect']['jumpToMessage']
   jumpToMessageStack: number[]
 }) {
+  const tx = useTranslationFunction()
+
   const countUnreadMessages = useUnreadCount(accountId, chat)
 
   let countToShow: string = countUnreadMessages.toString()
@@ -1000,6 +1002,11 @@ function JumpDownButton({
             'counter',
             countToShow.length === 3 && 'counter-3digits'
           )}
+          // Even though this is not focusable as of the time of writing,
+          // let's still apply label, for future-proofing.
+          aria-label={tx('chat_n_new_messages', String(countUnreadMessages), {
+            quantity: countUnreadMessages,
+          })}
           style={countUnreadMessages === 0 ? { visibility: 'hidden' } : {}}
         >
           {countToShow}